academic_service/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
waitress==3.0.0
Werkzeug==3.0.3
